+++ backend/backend/Models/SeasonEventScoreRules.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public float TimeMultiplier { get; set; }
         public float DistanceMultiplier { get; set; }
-        public int[] AvailableBonuses { get; set; }
+        public float[] AvailableBonuses { get; set; }
         public float UnfinishedSectorPenaltyPoints { get; set; }
         public bool TheMoreTheBetter { get; set; }
